automatically apply OneVsAll in CLinear (renamed from CLinearBinaryClassifierBuilder)
