+++ NeoML/docs/ru/Tutorial/News20Classification.md
@@ -11,8 +11,8 @@
 
 В этом примере мы используем библиотеку **NeoML** для обучения модели, решающей задачу многоклассовой классификации на классическом датасете [News20](https://archive.ics.uci.edu/ml/datasets/Twenty+Newsgroups).
 
-Для решения задачи используем комбинацию из двух методов: [линейного бинарного классификатора](../API/ClassificationAndRegression/Linear.md)
-и метода ["один против всех"](../API/ClassificationAndRegression/OneVersusAll.md).
+Для решения задачи используем комбинацию из двух методов: [линейного классификатора](../API/ClassificationAndRegression/Linear.md)
+и метода ["один против всех"](../API/ClassificationAndRegression/OneVersusAll.md), который включается неявно в случае, когда количество классов превышает 2.
 
 ## Подготовка входных данных
 
@@ -34,16 +34,13 @@ testArchive >> testData;
 
 ## Обучение классификатора
 
-Классификатор "один против всех" использует заданный бинарный классификатор для того, чтобы на каждый класс обучить модель, определяющую вероятность принадлежности объекта к этому классу. Отнесение объекта к одному из классов затем определяется путём голосования моделей.
+Метод "один против всех" использует заданный классификатор для того, чтобы на каждый класс обучить модель, определяющую вероятность принадлежности объекта к этому классу. Отнесение объекта к одному из классов затем определяется путём голосования моделей.
 
-1. Создадим линейный бинарный классификатор при помощи класса `CLinearBinaryClassifierBuilder`. В качестве функции потерь используем логистическую регрессию (константа `EF_LogReg`).
-2. Создадим классификатор `COneVersusAll`, передав только что настроенный бинарный классификатор на вход конструктору.
-3. Вызовем метод `Train`, который получает на вход обучающую выборку `trainData`, подготовленную на предыдущем шаге. Этот метод обучит модель и вернёт её в виде объекта, реализующего интерфейс `IModel`.
+1. Создадим линейный классификатор при помощи класса `CLinear` (`COneVersusAll` будет использован неявно). В качестве функции потерь используем логистическую регрессию (константа `EF_LogReg`).
+2. Вызовем метод `Train`, который получает на вход обучающую выборку `trainData`, подготовленную на предыдущем шаге. Этот метод обучит модель и вернёт её в виде объекта, реализующего интерфейс `IModel`.
 
 ```c++
-CLinearBinaryClassifierBuilder linear( EF_LogReg );
-	
-COneVersusAll oneVersusAll( linear );
+CLinear linear( EF_LogReg );
 CPtr<IModel> model = oneVersusAll.Train( *trainData );
 ```
