+++ NeoML/src/TraditionalML/Linear.cpp
@@ -16,7 +16,8 @@ limitations under the License.
 #include <common.h>
 #pragma hdrstop
 
-#include <NeoML/TraditionalML/LinearBinaryClassifierBuilder.h>
+#include <NeoML/TraditionalML/Linear.h>
+#include <NeoML/TraditionalML/OneVersusAll.h>
 #include <NeoML/TraditionalML/TrustRegionNewtonOptimizer.h>
 #include <LinearBinaryModel.h>
 #include <NeoML/TraditionalML/PlattScalling.h>
@@ -27,7 +28,7 @@ ILinearBinaryModel::~ILinearBinaryModel()
 {}
 
 // Normalizes the error weight
-static double normalizeErrorWeight( const CLinearBinaryClassifierBuilder::CParams& param, const IProblem& trainingClassificationData )
+static double normalizeErrorWeight( const CLinear::CParams& param, const IProblem& trainingClassificationData )
 {
 	double totalWeight = 0;
 	const int vectorCount = trainingClassificationData.GetVectorCount();
@@ -38,7 +39,7 @@ static double normalizeErrorWeight( const CLinearBinaryClassifierBuilder::CParam
 }
 
 // Normalizes the error weight
-static double normalizeErrorWeight( const CLinearBinaryClassifierBuilder::CParams& param, const IRegressionProblem& problem )
+static double normalizeErrorWeight( const CLinear::CParams& param, const IRegressionProblem& problem )
 {
 	double totalWeight = 0;
 	const int vectorCount = problem.GetVectorCount();
@@ -49,7 +50,7 @@ static double normalizeErrorWeight( const CLinearBinaryClassifierBuilder::CParam
 }
 
 // Creates a class for the loss function
-static CFunctionWithHessian* createOptimizedFunction( const CLinearBinaryClassifierBuilder::CParams& param, const IProblem& trainingClassificationData )
+static CFunctionWithHessian* createOptimizedFunction( const CLinear::CParams& param, const IProblem& trainingClassificationData )
 {
 	static_assert( EF_Count == 4, "EF_Count != 4" );
 
@@ -73,21 +74,21 @@ static CFunctionWithHessian* createOptimizedFunction( const CLinearBinaryClassif
 
 //---------------------------------------------------------------------------------------------------------
 
-CLinearBinaryClassifierBuilder::CLinearBinaryClassifierBuilder( const CParams& _params ) :
+CLinear::CLinear( const CParams& _params ) :
 	params( _params ),
 	log( 0 ),
 	function( 0 )
 {
 }
 
-CLinearBinaryClassifierBuilder::~CLinearBinaryClassifierBuilder()
+CLinear::~CLinear()
 {
 	if( function != 0 ) {
 		delete function;
 	}
 }
 
-CPtr<IRegressionModel> CLinearBinaryClassifierBuilder::TrainRegression( const IRegressionProblem& problem )
+CPtr<IRegressionModel> CLinear::TrainRegression( const IRegressionProblem& problem )
 {
 	if( function != 0 ) {
 		delete function; // delete the old loss function
@@ -108,8 +109,12 @@ CPtr<IRegressionModel> CLinearBinaryClassifierBuilder::TrainRegression( const IR
 	return FINE_DEBUG_NEW CLinearBinaryModel( plane, sigmoidCoefficients );
 }
 
-CPtr<IModel> CLinearBinaryClassifierBuilder::Train( const IProblem& trainingClassificationData )
+CPtr<IModel> CLinear::Train( const IProblem& trainingClassificationData )
 {
+	if( trainingClassificationData.GetClassCount() > 2 ) {
+		return COneVersusAll( *this ).Train( trainingClassificationData );
+	}
+
 	if( function != 0 ) {
 		delete function; // delete the old loss function
 	}
